sonarcloud feedback and PR feedback, light refactoring
